Improve DAOs and entities

+++ app/src/main/java/org/lagonette/android/room/entity/Partner.java
@@ -54,7 +54,7 @@ public class Partner {
     public String shortDescription;
 
     @ColumnInfo(name = "main_category")
-    public int mainCategory;
+    public long mainCategory;
 
     @Embedded
     public Address address;
